Fixup build and a few other things
This was not properly created to begin with so I am applying generous
amounts of duct tape and good intentions until we can safely re-build
this from scratch in the storedog folder.

I centralized the logging configuration into the lograge initializer and
tweaked the development environment to supress the cache hit debug
messages for now.