Fix a misunderstanding of context
This is what I get for reading v1 docs instead of the v2 docs and
getting confused. Context is really the folder path where the Dockerfile
lives. The file key is for the name of the Dockerfile at that directory
or if it lives in another directory. To simplify things, we're going
with the default behavior as it would be to build via docker locally
where you provide the context where the Dockerfile lives and it figures
out the rest.