Aha! We needed to make this a proc
This was a difficult bug to track down. The short answer here is that
Semantic Logger's Rack logging adapter will either do a call on a Proc
within the log tags or a call on any symbol against the incoming Rack
request object. It was difficult to understand this subtle difference,
but it solved the problem. My big clue was knowing the datadog tracer
was always supposed to return some kind of string, not nil.

Fixes #110