Flatten out payload
This is a pattern in the new semantic_logger library where a lot of
relevant log facets are being wrapped in a payload key inside the JSON
log payload which makes it difficult to parse out of the box. By
flattening the payload key back into the main structure we can get the
default pipeline to work out of the box for us.